+++ src/main/java/de/alpharogroup/db/entity/activatable/ActivatableEntity.java
@@ -58,6 +58,9 @@ public abstract class ActivatableEntity<PK extends Serializable> extends BaseEnt
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;
 
+	/** The Constant for the column name 'active'. */
+	public static final String COLUMN_NAME_ACTIVE = "active";
+
 	/** The attribute active, if true this entity is active. */
 	@Column(name = "active")
 	private boolean active;