+++ src/main/java/de/alpharogroup/db/entity/delete/Deletion.java
@@ -78,6 +78,12 @@ public class Deletion<PK extends Serializable, T, U> extends BaseEntity<PK>
 	/** The Constant serialVersionUID. */
 	private static final long serialVersionUID = 1L;
 
+	/** The Constant for the column name 'deleted'. */
+	public static final String COLUMN_NAME_DELETED = "deleted";
+
+	/** The Constant for the column name 'deletedBy'. */
+	public static final String COLUMN_NAME_DELETED_BY = "deletedBy";
+
 	/** The date and time when the entity that owns this entity was deleted. */
 	private T deleted;