+++ src/main/java/de/alpharogroup/db/entity/pair/KeyValueOwnerEntity.java
@@ -64,6 +64,15 @@ public abstract class KeyValueOwnerEntity<PK extends Serializable, O, K, V> exte
 	/** The Constant serialVersionUID. */
 	private static final long serialVersionUID = 1L;
 
+	/** The Constant for the column name 'key'. */
+	public static final String COLUMN_NAME_KEY = "key";
+
+	/** The Constant for the column name 'owner'. */
+	public static final String COLUMN_NAME_OWNER = "owner";
+
+	/** The Constant for the column name 'value'. */
+	public static final String COLUMN_NAME_VALUE = "value";
+
 	/** The key of this key-value pair. */
 	@Column(name = "key", nullable = false)
 	K key;