+++ src/main/java/de/alpharogroup/db/entity/text/TextEntity.java
@@ -57,6 +57,9 @@ public abstract class TextEntity<PK extends Serializable> extends BaseEntity<PK>
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;
 
+	/** The Constant for the column name 'text'. */
+	public static final String COLUMN_NAME_TEXT = "text";
+
 	/** The name. */
 	@Column(unique = false, name = "text", columnDefinition = "TEXT")
 	String text;