+++ src/main/java/de/alpharogroup/db/entity/uniqueable/UniqueableEntity.java
@@ -62,6 +62,9 @@ public abstract class UniqueableEntity<PK extends Serializable>
 	/** The serialVersionUID. */
 	private static final long serialVersionUID = 1L;
 
+	/** The Constant for the column name 'uuid'. */
+	public static final String COLUMN_NAME_UUID = "uuid";
+
 	/** The technical primary key. */
 	@Id
 	@EqualsAndHashCode.Include