+++ src/main/java/de/alpharogroup/db/entity/validation/ValidatableEntity.java
@@ -60,6 +60,12 @@ public class ValidatableEntity<PK extends Serializable, T> extends BaseEntity<PK
 	/** The serial Version UID. */
 	private static final long serialVersionUID = 1L;
 
+	/** The Constant for the column name 'validFrom'. */
+	public static final String COLUMN_NAME_VALID_FROM = "validFrom";
+
+	/** The Constant for the column name 'validTill'. */
+	public static final String COLUMN_NAME_VALID_TILL = "validTill";
+
 	/** The valid from date for validation. */
 	private T validFrom;