+++ CHANGELOG.md
@@ -6,7 +6,12 @@ Version 1.6-SNAPSHOT
 
 ADDED:
  
-- new constant for the name of columns
+- new constants for all relevant columns created
+- created new class VersionableEntity that replaces the class VersionableBaseEntity
+
+CHANGED:
+
+- VersionableBaseEntity tagged as deprecated
 
 Version 1.5
 -------------